Alice's up-B is a portal now, and the recovery is the portal rather than an arc
Jon's move, end to end: the technique, the adapter, and a real authored fighter.
Up-B opens an aperture under her and its partner 320px above; falling into one
leaves the other. ⛔ No recovery behaviour was written — `ambition_portal2d` owns
apertures, linking and transit, and this places two and counts down.

⭐ THE CRATE'S OWN PAIRING RULE, NOT A COPY OF IT. My first draft computed the
partner as `low ^ 1`, which is the rule `PortalChannelColor` documents — so I
deleted mine and called `PortalChannel::partner()`. A pairing rule with two homes
drifts apart the day the channel space changes shape. ⛔ And indices 8+ only,
refused at authoring: `0..=7` overlap the eight NAMED authored pairs, so a move
quietly taking one would link its portals to whatever a room authored in that
colour — which reads in play as the recovery teleporting you somewhere absurd.

⛔ THE IMPULSE IS GONE FROM `elliptic_curve`, and that is the one behavioural
change. A `-760` `Set` alongside the pair would make the portal decorative and
the arc the real recovery, which is the opposite of the move. The swing, the
rune, the landing lag and the timing are untouched — the pair opens on the beat
the impulse used to fire.

⭐⭐ AND THE ASSERTION THAT CATCHES THAT WAS WRONG UNTIL A POISON DID NOT FIRE.
It looked for an `Effect` whose key contained "impulse"; an impulse is
`MoveEventKind::Impulse`, its own variant, so the check could never match and
restoring the impulse left the test GREEN. ⇒ Third time today a non-firing poison
found a weak assertion rather than correct code, which is becoming the most
reliable review I have.

⛔ A SECOND WEAK GUARD, SAME WAY. "An expired pair closes at both ends" could not
distinguish my pair-index sweep from "close whatever expired", because both
apertures share a lifetime and tick in lockstep. The property is only observable
when their clocks differ, so the test now constructs that directly — and the
narrowing poison fires.

⭐ THE LIFETIME STAYS AUTHORED. `close_on_transit: false` makes this a ROUTE
anyone may take, including whoever is chasing her, rather than a hatch that shuts
behind her. That is the interesting version and it is a one-word change if it
plays badly. ⚠ 320px, 2.5s and the aperture size are knobs against a sprite.

⇒ ROLLBACK PAID: `smash.move_placed_portal` registered `component-clone` + probed
beside `smash.live_bomb`, whose fuse doc gives the reason verbatim — a rewind
that restores the aperture without its clock gives the resimulated timeline a
route that closes at a different moment.
⛔⛔ AND THE BASELINE'S VERSION LINE WAS STALE FROM MY OWN EARLIER COMMIT:
`a103a417e` bumped `GGRS_ROLLBACK_SCHEMA_VERSION` to 152 and did not update
`ggrs-rollback-schema-v151` in the recorded dump, so that ledger has been red
since. Both rows corrected here; the wire baseline test passes.

+++ crates/ambition_characters/src/smash_portal.rs
@@ -61,12 +61,15 @@ pub struct PortalPairParams {
     /// possible test of whether the placement seam takes an orientation at all.
     #[serde(default)]
     pub tilt_degrees: f32,
-    /// Which indexed channel pair to use. The partner is `index ^ 1`.
+    /// Which indexed channel pair to use. ⭐ The PARTNER is not computed here —
+    /// `PortalChannel::partner()` owns that rule, and a second copy of "the
+    /// partner is `n ^ 1`" is a pairing rule with two homes.
     ///
     /// ⛔ INDICES `8..` ONLY, which the colour table asks for in place: `0..=7`
     /// overlap the eight NAMED pairs in index space, and a move that quietly
-    /// took one would fight whatever authored that colour in the room.
-    pub channel_index: u16,
+    /// took one would fight whatever authored that colour in the room. ⓘ The
+    /// ceiling is the type's: `Indexed(u8)`, so 128 distinct pairs.
+    pub channel_index: u8,
 }
 
 /// Author a portal pair onto `spec`, opening at `at_s`.
